Keeping model-specific components in the model path, not generalized
Address review: model-specific quantization handling belongs in the
qwen3_5 model folder, with utils.py only dispatching by model_type
(mirroring the DeepSeek V4 pattern). Runtime behavior is unchanged.

- Move dequantize_fp8_weights (and FP8_BLOCK_SIZE) from
  models/qwen3_5/weights.py into models/qwen3_5/language.py; delete
  weights.py and repoint its importers (qwen3_5.py, the MTP drafter,
  and the model test).
- Move the ModelOpt MXFP4 transform (_f32_to_e8m0,
  transform_modelopt_mxfp4_weights) out of utils.py into
  models/qwen3_5/language.py; utils.py now imports and dispatches it
  only for model_type == 'qwen3_5'.
- Revert the unrelated ar.py prompt-token-counter change.

FP8 checkpoints are still dequantized to bf16 in the model's sanitize()
(unchanged); the model-owned entry points keep the format-specific
implementation in the model folder.
