+++ mlx_vlm/models/qwen3_5/language.py
@@ -1,5 +1,5 @@
 from functools import lru_cache, partial
-from typing import Any, List, Optional
+from typing import Any, Dict, List, Optional, Tuple
 
 import mlx.core as mx
 import mlx.nn as nn
@@ -20,6 +20,111 @@ from .gated_delta import (
     gated_delta_update,
     gated_delta_update_with_states,
 )
+
+FP8_BLOCK_SIZE = 128
+
+
+def dequantize_fp8_weights(
+    weights: Dict[str, mx.array], dtype=mx.bfloat16
+) -> Dict[str, mx.array]:
+    weights = dict(weights)
+    scale_suffix = ".weight_scale_inv"
+
+    for scale_key in [key for key in weights if key.endswith(scale_suffix)]:
+        weight_key = scale_key.removesuffix("_scale_inv")
+        if weight_key not in weights:
+            raise ValueError(f"Missing FP8 weight for {scale_key}.")
+
+        weight = weights[weight_key]
+        scale = weights.pop(scale_key)
+        if weight.ndim != 2:
+            raise ValueError(f"Expected a matrix for {weight_key}, got {weight.shape}.")
+        rows, columns = weight.shape
+        expected_scale_shape = (
+            (rows + FP8_BLOCK_SIZE - 1) // FP8_BLOCK_SIZE,
+            (columns + FP8_BLOCK_SIZE - 1) // FP8_BLOCK_SIZE,
+        )
+        if scale.shape != expected_scale_shape:
+            raise ValueError(
+                f"Invalid FP8 scale shape for {weight_key}: expected "
+                f"{expected_scale_shape}, got {scale.shape}."
+            )
+
+        padded_rows = expected_scale_shape[0] * FP8_BLOCK_SIZE
+        padded_columns = expected_scale_shape[1] * FP8_BLOCK_SIZE
+        decoded = mx.from_fp8(weight, dtype=mx.float32)
+        decoded = mx.pad(
+            decoded,
+            ((0, padded_rows - rows), (0, padded_columns - columns)),
+        )
+        decoded = decoded.reshape(
+            expected_scale_shape[0],
+            FP8_BLOCK_SIZE,
+            expected_scale_shape[1],
+            FP8_BLOCK_SIZE,
+        )
+        decoded = decoded * scale.astype(mx.float32)[:, None, :, None]
+        decoded = decoded.reshape(padded_rows, padded_columns)[:rows, :columns]
+        decoded = decoded.astype(dtype)
+        # Bound peak memory while converting a full sharded checkpoint.
+        mx.eval(decoded)
+        weights[weight_key] = decoded
+
+    return weights
+
+
+def _f32_to_e8m0(x: mx.array) -> mx.array:
+    """Encode non-negative scales to the E8M0 bytes used by MLX MXFP4."""
+    x = x.astype(mx.float32)
+    finite = mx.isfinite(x)
+    positive = x > 0
+    safe = mx.where(finite & positive, x, 1.0)
+    exponent = mx.clip(mx.round(mx.log2(safe)), -127, 127) + 127
+    exponent = mx.where(finite, mx.where(positive, exponent, 0), 255)
+    return exponent.astype(mx.uint8)
+
+
+def transform_modelopt_mxfp4_weights(
+    weights: Dict[str, mx.array],
+    quantization_config: Optional[Dict[str, Any]],
+) -> Tuple[Dict[str, mx.array], Optional[Dict[str, Any]]]:
+    if quantization_config is None:
+        return weights, None
+    if quantization_config.get("quant_method") != "mxfp4":
+        return weights, None
+
+    scale_suffix = ".weight_scale"
+    prefixes = {
+        key[: -len(scale_suffix)] for key in weights if key.endswith(scale_suffix)
+    }
+    if not prefixes:
+        return weights, None
+    if quantization_config.get("micro_block", 32) != 32:
+        raise ValueError("MLX MXFP4 requires a micro block size of 32.")
+
+    transformed = {}
+    for key, value in weights.items():
+        if key.endswith(scale_suffix):
+            prefix = key[: -len(scale_suffix)]
+            weight_key = f"{prefix}.weight"
+            if weight_key not in weights:
+                raise ValueError(f"Missing MXFP4 weight for {key}.")
+            weight = weights[weight_key]
+            if weight.dtype != mx.uint8 or weight.ndim != 2 or value.ndim != 2:
+                raise ValueError(f"Invalid ModelOpt MXFP4 tensors for {prefix}.")
+            if (
+                weight.shape[0] != value.shape[0]
+                or weight.shape[1] != 16 * value.shape[1]
+            ):
+                raise ValueError(f"Invalid ModelOpt MXFP4 scale shape for {prefix}.")
+            transformed[weight_key] = weight.view(mx.uint32)
+            transformed[f"{prefix}.scales"] = _f32_to_e8m0(value)
+        elif key.endswith(".weight") and key[: -len(".weight")] in prefixes:
+            continue
+        else:
+            transformed[key] = value
+
+    return transformed, {"group_size": 32, "bits": 4, "mode": "mxfp4"}
 
 
 class Qwen3_5RotaryEmbedding(MRoPERotaryEmbedding):
